% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPopulationSimOutput.R
\name{ReadPopulationSimOutput}
\alias{ReadPopulationSimOutput}
\title{Main module function to reads PopulationSim output}
\usage{
ReadPopulationSimOutput(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module along with:
LENGTH: A named integer vector having a single named element, "Household",
which identifies the length (number of rows) of the Household table to be
created in the datastore.
SIZE: A named integer vector having two elements. The first element, "Azone",
identifies the size of the longest Azone name. The second element, "HhId",
identifies the size of the longest HhId.
}
\description{
\code{ReadPopulationSimOutput} reads and stores the synthetic households and
persons produced by running PopulationSim in the VisionEval framework.
}
